backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.6,<5.0
psycopg2-binary>=2.9,<3.0
openpyxl>=3.1,<4.0
gunicorn>=23.0,<24.0
Pillow>=11.0,<12.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=6.0,<7.0
factory-boy>=3.3,<4.0
